refactored basictracer/tracer.go to tracer.go
