+++ basictracer/util.go
@@ -12,13 +12,13 @@ var (
 	seededIDLock sync.Mutex
 )
 
-func randomID() uint64 {
+func RandomID() uint64 {
 	seededIDLock.Lock()
 	defer seededIDLock.Unlock()
 	return uint64(seededIDGen.Int63())
 }
 
-func randomID2() (uint64, uint64) {
+func RandomID2() (uint64, uint64) {
 	seededIDLock.Lock()
 	defer seededIDLock.Unlock()
 	return uint64(seededIDGen.Int63()), uint64(seededIDGen.Int63())
