+++ propagation_ls.go
@@ -1,9 +1,10 @@
-package basictracer
+package lightstep
 
 import (
 	"encoding/base64"
 
 	"github.com/gogo/protobuf/proto"
+	"github.com/lightstep/lightstep-tracer-go/basictracer"
 	lightstep "github.com/lightstep/lightstep-tracer-go/lightsteppb"
 	opentracing "github.com/opentracing/opentracing-go"
 )
@@ -22,7 +23,7 @@ func (_ lightstepBinaryPropagator) Inject(
 	spanContext opentracing.SpanContext,
 	opaqueCarrier interface{},
 ) error {
-	sc, ok := spanContext.(SpanContext)
+	sc, ok := spanContext.(basictracer.SpanContext)
 	if !ok {
 		return opentracing.ErrInvalidSpanContext
 	}
@@ -92,7 +93,7 @@ func (_ lightstepBinaryPropagator) Extract(
 		return nil, opentracing.ErrInvalidCarrier
 	}
 
-	return SpanContext{
+	return basictracer.SpanContext{
 		TraceID: pb.BasicCtx.TraceId,
 		SpanID:  pb.BasicCtx.SpanId,
 		Baggage: pb.BasicCtx.BaggageItems,
